End of day commit 30/01/23
Code written for extracting the zip file from 'tmp/landing/*.zip' to 'tmp/Steven/landing'. Further code has been written to read from 'tmp/Steven/landing' to 'tmp/Steven/Bronze/Relevant Folder/Delta File'. The code will extract all the zip files located in the landing folder, and will also create individual delta folders in the 'tmp/Steven/Landing/'. Each contains the relevant delta files, with each column set to string type.

+++ Testing Notebook.py
@@ -80,7 +80,7 @@ rider_silver_schema = StructType ([
 payment_silver_schema = StructType ([
     StructField("payment_id", IntegerType(), True),
     StructField("date", DateType(), True),
-    StructField("amount", DecimalType(), True),
+    StructField("amount", FloatType(), True),
     StructField("rider_id", IntegerType(), True),
 ])
